Update TransH regularizer
