+++ src/pykeen/regularizers.py
@@ -198,34 +198,33 @@ class TransHRegularizer(Regularizer):
 
     def __init__(
         self,
+        entity_embeddings: nn.Parameter,
+        normal_vector_embeddings: nn.Parameter,
+        relation_embeddings: nn.Parameter,
         weight: float = 0.05,
         epsilon: float = 1e-5,
-        parameters: Optional[Iterable[nn.Parameter]] = None,
     ):
         # The regularization in TransH enforces the defined soft constraints that should computed only for every batch.
         # Therefore, apply_only_once is always set to True.
-        super().__init__(weight=weight, apply_only_once=True, parameters=parameters)
+        super().__init__(weight=weight, apply_only_once=True, parameters=[])
+        self.normal_vector_embeddings = normal_vector_embeddings
+        self.relation_embeddings = relation_embeddings
+        self.entity_embeddings = entity_embeddings
         self.epsilon = epsilon
 
     def forward(self, x: torch.FloatTensor) -> torch.FloatTensor:  # noqa: D102
         raise NotImplementedError('TransH regularizer is order-sensitive!')
 
-    def update(self, *tensors: torch.FloatTensor) -> None:  # noqa: D102
-        if len(tensors) != 3:
-            raise KeyError('Expects exactly three tensors')
-        if self.apply_only_once and self.updated:
-            return
-        entity_embeddings, normal_vector_embeddings, relation_embeddings = tensors
+    def pop_regularization_term(self, *tensors: torch.FloatTensor) -> torch.FloatTensor:  # noqa: D102
         # Entity soft constraint
-        self.regularization_term += torch.sum(functional.relu(torch.norm(entity_embeddings, dim=-1) ** 2 - 1.0))
+        self.regularization_term += torch.sum(functional.relu(torch.norm(self.entity_embeddings, dim=-1) ** 2 - 1.0))
 
         # Orthogonality soft constraint
-        d_r_n = functional.normalize(relation_embeddings, dim=-1)
+        d_r_n = functional.normalize(self.relation_embeddings, dim=-1)
         self.regularization_term += torch.sum(
-            functional.relu(torch.sum((normal_vector_embeddings * d_r_n) ** 2, dim=-1) - self.epsilon),
+            functional.relu(torch.sum((self.normal_vector_embeddings * d_r_n) ** 2, dim=-1) - self.epsilon),
         )
-
-        self.updated = True
+        return super().pop_regularization_term()
 
 
 class CombinedRegularizer(Regularizer):
